add ModuleTableItem file
