+++ app/src/assets/localization/en/protocol_setup.json
@@ -101,6 +101,8 @@
   "heater_shaker_extra_attention": "Use latch controls for easy placement of labware.",
   "heater_shaker_labware_list_view": "To add labware, use the toggle to control the latch",
   "home_stacker": "Home stacker",
+  "home_stacker_warning_title": "Close robot and stacker doors",
+  "home_stacker_warning_description": "For your safety, ensure doors are closed prior to homing the stacker.",
   "how_offset_data_works": "How labware offsets work",
   "individiual_well_volume": "Individual well volume",
   "initial_liquids_num": "{{count}} initial liquid",
